# docs/tech/api/openapi.yaml
openapi: 3.0.3
info:
  title: Pereobuyka API
  description: |
    HTTP API ядра «Переобуйка». Сценарии — docs/vision.md §4.
    LLM не подменяет факты; контекст подмешивается на сервере (vision §6).
  version: 1.0.0
servers:
  - url: /
...
paths:
  /api/v1/auth/telegram:
    post:
      tags: [Auth]
      summary: Вход / регистрация через Telegram (upsert User)
      operationId: authTelegram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramAuthRequest"
      responses:
        "200":
          description: OK — существующий пользователь вошёл
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "201":
          description: Created — новый пользователь зарегистрирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/auth/web:
    post:
      tags: [Auth]
      summary: Вход клиента в веб по telegram_username (MVP)
      operationId: authWeb
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebAuthRequest"
      responses:
        "200":
          description: OK — пользователь найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "201":
          description: Created — создан новый клиент
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/me:
    get:
      tags: [Client]
      summary: Текущий пользователь
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/me/appointments:
    get:
      tags: [Client]
      summary: Мои записи
      operationId: listMyAppointments
      security:
        - BearerAuth: []
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/AppointmentStatus"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/me/appointments/{appointment_id}:
    parameters:
      - $ref: "#/components/parameters/AppointmentId"
    patch:
      tags: [Client]
      summary: Обновить запись (в т.ч. отменить)
      operationId: patchMyAppointment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentPatchRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/DomainError"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/me/visits:
    get:
      tags: [Client]
      summary: История визитов
      operationId: listMyVisits
      security:
        - BearerAuth: []
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/me/visits/{visit_id}/service-rating:
    post:
      tags: [Client]
      summary: Оценка сервиса после визита
      operationId: postServiceRating
      security:
        - BearerAuth: []
      parameters:
        - name: visit_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRatingBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visit"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/me/bonus-account:
    get:
      tags: [Client]
      summary: Бонусный счёт
      operationId: getMyBonusAccount
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BonusAccount"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/me/bonus-transactions:
    get:
      tags: [Client]
      summary: История бонусных операций
      operationId: listMyBonusTransactions
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BonusTransactionListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    AppointmentId:
      name: appointment_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Нет или невалидный токен
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Ошибка валидации тела или параметров
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    DomainError:
      description: Доменная ошибка (конфликт, нарушение бизнес-правила)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    UserRole:
      type: string
      enum: [client, admin]
    UserSource:
      type: string
      enum: [telegram, web]
    User:
      type: object
      required: [id, name, role, registered_at, source]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        telegram_id:
          type: integer
          format: int64
          nullable: true
        telegram_username:
          type: string
          nullable: true
        registered_at:
          type: string
          format: date-time
        source:
          $ref: "#/components/schemas/UserSource"
    TelegramAuthRequest:
      type: object
      required: [telegram_id, name]
      properties:
        telegram_id:
          type: integer
          format: int64
        name:
          type: string
        phone:
          type: string
    WebAuthRequest:
      type: object
      required: [telegram_username]
      properties:
        telegram_username:
          type: string
          minLength: 1
          maxLength: 64
        name:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
    TokenResponse:
      type: object
      required: [access_token, token_type, user]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
        user:
          $ref: "#/components/schemas/User"

    # ── Services ───────────────────────────────────────────────────────────────
...
    AppointmentStatus:
      type: string
      enum: [scheduled, completed, cancelled]
    AppointmentSource:
      type: string
      enum: [llm, telegram_bot, web, admin]
    ServiceLineItem:
      type: object
      description: Позиция услуги в записи или визите
      required: [service_id, quantity]
      properties:
        service_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          default: 1
...
    AppointmentPatchRequest:
      type: object
      properties:
        status:
          type: string
          enum: [cancelled]
          description: Клиент может перевести запись только в статус cancelled
    Appointment:
      type: object
      required:
        - id
        - user_id
        - starts_at
        - ends_at
        - total_price
        - status
        - created_at
        - service_items
        - source
        - discount_percent
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        total_price:
          type: string
          description: "Расчётная стоимость на момент записи, десятичная строка"
          example: "4300.00"
        status:
          $ref: "#/components/schemas/AppointmentStatus"
        created_at:
          type: string
          format: date-time
        service_items:
          type: array
          items:
            $ref: "#/components/schemas/ServiceLineItem"
        source:
          $ref: "#/components/schemas/AppointmentSource"
        discount_percent:
          type: integer
          minimum: 0
          maximum: 100
    AppointmentListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer
          description: Общее число записей без учёта пагинации
...
    Visit:
      type: object
      required:
        - id
        - appointment_id
        - total_amount
        - bonus_spent
        - bonus_earned
        - confirmed_at
        - confirmed_by_user_id
        - lines
      properties:
        id:
          type: string
          format: uuid
        appointment_id:
          type: string
          format: uuid
        total_amount:
          type: string
          description: "Итоговая сумма оплаты, десятичная строка"
          example: "4300.00"
        bonus_spent:
          type: integer
        bonus_earned:
          type: integer
          description: Начислено бонусов по итогам визита (рассчитывает сервер)
        confirmed_at:
          type: string
          format: date-time
        confirmed_by_user_id:
          type: string
          format: uuid
        lines:
          type: array
          items:
            $ref: "#/components/schemas/ServiceLineItem"
        client_rating_stars:
          type: integer
          nullable: true
          minimum: 1
          maximum: 5
        client_rating_comment:
          type: string
          nullable: true
        service_rating_stars:
          type: integer
          nullable: true
          minimum: 1
          maximum: 5
        service_rating_comment:
          type: string
          nullable: true
    VisitListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Visit"
        total:
          type: integer
          description: Общее число визитов без учёта пагинации
...
    BonusAccount:
      type: object
      required: [id, user_id, balance]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        balance:
          type: integer
    BonusTransactionType:
      type: string
      enum: [earn, spend, adjust]
    BonusTransaction:
      type: object
      required: [id, type, amount, created_at]
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/BonusTransactionType"
        amount:
          type: integer
          description: "Положительное — начисление, отрицательное — списание"
        visit_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        comment:
          type: string
          nullable: true
    BonusTransactionListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/BonusTransaction"
        total:
          type: integer
...
    ServiceRatingBody:
      type: object
      required: [stars]
      properties:
        stars:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true